Move funcs from `vpn` to `util/netutil`, implement some
- `DefaultNetworkInterface`, `LocalNetworkInterfaceIPs`, `NetworkInterfaceIPs` were moved from `vpn` to `util/netutil`;
- `DefaultNetworkInterface` was implemented for MacOS.
